\name{monitoring_overview}

\title{monitoring_overview}
\usage{
monitoring_overview(file_name)
}
\description{
Gives graphic showing all monitoring since 2006 based on the .csv file.
}

\format{

\itemize{
\item \code{file_name}: .csv file containing the data for the analysis.   
 
}

}
\examples{
monitoring_overview(file_name = "Extract_otago.csv")

}



}
