\name{sloop_folder_structure}
\alias{sloop_folder_structure}

\title{sloop_folder_structure}
\usage{
sloop_folder_structure(folder)
}
\description{
Lists all the files contained within the named folder. Only folders within the project are considered
}

\format{

\itemize{
\item \code{folder}: the folder within the project to be mapped.   
 
}
}

\examples{
sloop_folder_structure(folder = "Sloop_robust_design")
}




